requirements.txt
actionlib==1.14.0
angles==1.9.13
ansitable==0.11.4
base_local_planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-tools==0.9.5
cfgv==3.4.0
colored==1.4.4
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.11.0
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.9
distro==1.9.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
empy==4.2
filelock==3.16.1
fonttools==4.38.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
identify==2.6.1
image-geometry==1.16.2
importlib-metadata==6.7.0
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.5
laser_geometry==1.6.7
matplotlib==3.5.3
message-filters==1.17.0
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
nodeenv==1.9.1
numpy==1.24.4
osrf-pycommon==2.0.2
packaging==24.0
pgraph-python==0.6.3
Pillow==9.5.0
platformdirs==4.3.6
pre-commit==3.5.0
progress==1.6
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.16
PyQt5-sip==12.13.0
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.8
roboticstoolbox-python==1.1.1
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rtb-data==1.0.1
rviz==1.14.25
scipy==1.10.1
sensor-msgs==1.13.1
shiboken2==5.15.2.1
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.6
spatialgeometry==1.1.0
spatialmath-python==1.1.14
srdfdom==0.6.4
swift-sim==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threaded==4.2.0
topic-tools==1.17.0
typing_extensions==4.7.1
urdfdom-py==0.4.6
virtualenv==20.30.0
websockets==13.1
xacro==1.14.19
zipp==3.15.0
